fix(redact): extract redact() to shared core utility; apply at MCP run_channel/run_clarify boundaries; redact experience event queries before write
